Implementada lógica de busca em profundidade

+++ ClassLibrary1/Model/QueuedURL.cs
@@ -9,9 +9,9 @@ namespace WebCrawler.Model
     internal class QueuedURL
     {
         public string URL { get; set; }
-        public int priority { get; set; }
+        public double priority { get; set; }
 
-        internal QueuedURL(string URL, int Priority)
+        internal QueuedURL(string URL, double Priority)
         {
             this.URL = URL;
             this.priority = Priority;
